+++ app/components/map/aircraft/MapAircraft.vue
@@ -292,7 +292,7 @@ const isSelfFlight = computed(() => props.aircraft?.cid === ownFlight.value?.cid
 const zoom = computed(() => mapStore.zoom);
 
 function checkForExpiredCoordinate() {
-    if (dataStore.vatsim.selfCoordinate.value && dataStore.vatsim.updateTime.value - dataStore.vatsim.selfCoordinate.value.date > 1000 * 5) {
+    if (dataStore.vatsim.selfCoordinate.value && dataStore.vatsim.localUpdateTime.value - dataStore.vatsim.selfCoordinate.value.date > 1000 * 5) {
         dataStore.vatsim.selfCoordinate.value = null;
         return [props.aircraft.longitude, props.aircraft.latitude];
     }
@@ -540,6 +540,7 @@ async function setPilotRoute(enabled: boolean) {
 
         dataStore.navigraphWaypoints.value[props.aircraft.cid.toString()] = {
             pilot: pilot.value,
+            coordinates: getCoordinates.value,
             full: typeof activeCurrentOverlay.value?.data?.fullRoute === 'boolean' ? activeCurrentOverlay.value?.data?.fullRoute : !!store.user?.settings.showFullRoute,
             calculatedArrival: dataStore.navigraphWaypoints.value[props.aircraft.cid.toString()]?.calculatedArrival,
             disableHoldings: store.localSettings.navigraphRouteAirportOverlay?.holds === false && !activeCurrentOverlay.value && !props.isHovered,