+++ app/utils/backend/worker/navigraph.ts
@@ -46,6 +46,8 @@ async function updateNavigraph() {
     try {
         updating = true;
         if (navigraphData.versions.current !== radarStorage.navigraph.current || navigraphData.versions.outdated !== radarStorage.navigraph.outdated || process.env.NODE_ENV === 'development') {
+            console.log('Update has started', radarStorage.navigraph.current, navigraphData.versions.current);
+            await unsetRedisSync('navigraph-ready');
             const current = await processDatabase(navigraphCurrentDb!, radarStorage.navigraph.current);
             const outdated = await processDatabase(navigraphOutdatedDb!, radarStorage.navigraph.outdated);
 
@@ -59,6 +61,7 @@ async function updateNavigraph() {
 
             setRedisSync('navigraph-ready', '1', 1000 * 60 * 60 * 24);
             defaultRedis.publish('update', 'navigraph-data');
+            console.log(`Cycle is ${ navigraphData.versions.current }`);
         }
         else {
             setRedisSync('navigraph-ready', '1', 1000 * 60 * 60 * 24);